+++ legacy_packages/unity-js-bridge/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @thirdweb-dev/unity-js-bridge
 
+## 0.6.40
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @thirdweb-dev/sdk@4.0.91
+  - @thirdweb-dev/wallets@2.5.31
+  - @thirdweb-dev/auth@4.1.89
+
+## 0.6.39
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @thirdweb-dev/sdk@4.0.90
+  - @thirdweb-dev/wallets@2.5.30
+  - @thirdweb-dev/auth@4.1.88
+
 ## 0.6.38
 
 ### Patch Changes